Data product last modification time

This data frame records the last modify time for each data product in this package.

data_summary

Format

A data frame with the following columns:

taxon_group: The taxa group that the location information can be used for. Note that some taxa groups may have the same 'plotID' but their latitude/longitude may differ slightly, which justifies the need of this column.

data_package_id: Identifier for the data generated by {ecocomDP}.

n_taxa: Number of species included. Note that we did not clean species anmes though we did remove records with species identified above genus level for well studied groups such as plant and fish. However, we keep all records regardless of their species identification levels for taxonomic groups that are hard to identifiy (e.g. macroinvertebrate, beetle).

n_sites: Number of NEON sites included.

sites: All site codes that have data, separated by |.

start_date: The earliest date that have records.

end_date: The latest date that have records.

data_package_title: Title of the dataset.

neon_ecocomdp_mapping_method: This is the ID that ecocomDP::read_data used to download and process the NEON data product with NEON's orignal ID specified in original_neon_data_product_id.

original_neon_data_product_id: The NEON data product ID. See neonUtilities:::table_types for all available data types and their data product IDs provided by NEON.

original_neon_data_version: NEON data release version.

original_neon_data_doi: Original NEON data doi.

r_object: The name of R objects for each taxonomic group. By calling the R objects, we can get the processed and standardized NEON organismal data for downstream diversity analysis.

variable_names: The variable names that represent "abundance" information.

units: The units of the variable_names.